backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic[email]==2.3.0
sqlalchemy==2.0.20
alembic==1.12.0
asyncpg==0.28.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.0.1
